+++ R/learner_RWeka_classif_IBk.R
@@ -35,8 +35,8 @@ LearnerClassifIBk = R6Class("LearnerClassifIBk",
     #' Creates a new instance of this [R6][R6::R6Class] class.
     initialize = function() {
       ps = ps(
-        subset = p_uty(tags = c("train", "pars")),
-        na.action = p_uty(tags = c("train", "pars")),
+        subset = p_uty(tags = "train"),
+        na.action = p_uty(tags = "train"),
         I = p_lgl(default = FALSE, tags = "train"),
         F = p_lgl(default = FALSE, tags = "train"),
         K = p_int(default = 1L, lower = 1L, tags = "train"),