+++ R/learner_mboost_regr_glmboost.R
@@ -73,13 +73,13 @@ LearnerRegrGLMBoost = R6Class("LearnerRegrGLMBoost",
 
       pars = self$param_set$get_values(tags = "train")
       pars_boost = pars[which(names(pars) %in%
-        methods::formalArgs(mboost::boost_control))]
+        formalArgs(mboost::boost_control))]
       pars_glmboost = pars[which(names(pars) %in%
-        methods::formalArgs(mboost::gamboost))]
+        formalArgs(mboost::gamboost))]
 
       if (self$param_set$values$family != "custom") {
         pars_family = pars[which(names(pars) %in%
-          methods::formalArgs(utils::getFromNamespace(
+          formalArgs(utils::getFromNamespace(
             pars_glmboost$family,
             asNamespace("mboost"))))]
       }