+++ R/learner_mboost_surv_gamboost.R
@@ -122,20 +122,20 @@ LearnerSurvGAMBoost = R6Class("LearnerSurvGAMBoost",
       family = switch(pars$family,
         coxph = mboost::CoxPH(),
         weibull = mlr3misc::invoke(mboost::Weibull,
-          .args = pars[names(pars) %in% methods::formalArgs(mboost::Weibull)]),
+          .args = pars[names(pars) %in% formalArgs(mboost::Weibull)]),
         loglog = mlr3misc::invoke(mboost::Loglog,
-          .args = pars[names(pars) %in% methods::formalArgs(mboost::Loglog)]),
+          .args = pars[names(pars) %in% formalArgs(mboost::Loglog)]),
         lognormal = mlr3misc::invoke(mboost::Lognormal,
-          .args = pars[names(pars) %in% methods::formalArgs(mboost::Lognormal)]),
+          .args = pars[names(pars) %in% formalArgs(mboost::Lognormal)]),
         gehan = mboost::Gehan(),
         cindex = mlr3misc::invoke(mboost::Cindex,
-          .args = pars[names(pars) %in% methods::formalArgs(mboost::Cindex)]),
+          .args = pars[names(pars) %in% formalArgs(mboost::Cindex)]),
         custom = pars$custom.family
       )
 
       # FIXME - until issue closes
-      pars = pars[!(names(pars) %in% methods::formalArgs(mboost::Weibull))]
-      pars = pars[!(names(pars) %in% methods::formalArgs(mboost::Cindex))]
+      pars = pars[!(names(pars) %in% formalArgs(mboost::Weibull))]
+      pars = pars[!(names(pars) %in% formalArgs(mboost::Cindex))]
       pars = pars[!(names(pars) %in% c("family", "custom.family"))]