+++ R/learner_np_dens_mixed.R
@@ -72,7 +72,7 @@ LearnerDensMixed = R6Class("LearnerDensMixed",
 
   private = list(
     .train = function(task) {
-      pars = self$param_set$get_values(tag = "train")
+      pars = self$param_set$get_values(tags = "train")
       data = task$data()[[1]]
 
       pdf = function(x) {} # nolint